Unify controllers for PromoRegistrations
