+++ app/controllers/publishers/promo_registrations_controller.rb
@@ -1,5 +1,29 @@
-class Publishers::PromoRegistrationsController < PublishersController
+class Publishers::PromoRegistrationsController < ApplicationController
+  include PromosHelper
+
+  before_action :authenticate_publisher!
+  before_action :require_publisher_promo_disabled, :require_promo_running, only: %i(create)
+
+  layout "promo_registrations", only: [:index, :create]
+
   GROUP_START_DATE = Date.new(2019, 10, 1)
+
+  def index
+    @publisher = current_publisher
+    @publisher_promo_status = @publisher.promo_status(promo_running?)
+    @promo_enabled_channels = @publisher.channels.joins(:promo_registration)
+  end
+
+  def create
+    @publisher = current_publisher
+    @publisher.promo_enabled_2018q1 = true
+    @publisher.save!
+    current_publisher.channels.find_each do |channel|
+      channel.register_channel_for_promo # Callee does a check
+    end
+    @promo_enabled_channels = @publisher.channels.joins(:promo_registration)
+    @publisher_has_verified_channel = @publisher.has_verified_channel?
+  end
 
   def for_referral_code
     promo_registration = user.promo_registrations.find_by(referral_code: params[:referral_code])
@@ -38,6 +62,14 @@ class Publishers::PromoRegistrationsController < PublishersController
   end
 
   private
+
+  def require_promo_running
+    redirect_to promo_registrations_path, action: "index" unless promo_running?
+  end
+
+  def require_publisher_promo_disabled
+    redirect_to promo_registrations_path, action: "index" if current_publisher.promo_enabled_2018q1
+  end
 
   # In October 2019 we had an instance where users could receive confirmations with the legacy $5 rate along with the new group rates.
   # We need to show users how many confirmations the received in different groups along with the previous rate.
