+++ app/views/publishers/promo_registrations/_activate.html.slim
