+++ app/views/publishers/promo_registrations/_activated_unverified.html.slim
