+++ app/views/publishers/promo_registrations/_activated_verified.html.slim
