+++ app/views/publishers/promo_registrations/_active.html.slim
