+++ app/views/publishers/promo_registrations/_over.html.slim
