+++ app/views/publishers/promo_registrations/_promo_footer.html.slim
