+++ app/views/publishers/promo_registrations/create.html.slim
