+++ app/views/publishers/promo_registrations/index.html.slim
